realpath: Use physical PWD
This was an inadvertent change from
cc632d6ae95bdf3ebde068358397666b6a1b3b68.

Because we used wgetcwd directly before, we always got the "physical"
resolved $PWD.

There's an argument to be made to use the logical $PWD here as well
but I prefer not to make changes lik that in a random commit without
good reason.