Disclose pager to screen height immediately
This removes that bit where we only show 4 rows at most at first,
instead we disclose up to the full terminal height.

This results in less pressing of tab to get the other results, and
better visibility of results.

However, it also means that it'll push the shell up to the top of the
terminal a lot.

So this is an experiment to see how that works in practice.

Fixes #2698